+++ App/app/src/main/java/com/example/testeandroidv2/domain/login/ErrorLogin.kt
@@ -1,7 +1,6 @@
-package com.example.testeandroidv2.model
+package com.example.testeandroidv2.domain.login
 
 
-import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
