+++ App/app/src/main/java/com/example/testeandroidv2/domain/login/LoginBodyResponse.kt
@@ -1,4 +1,4 @@
-package com.example.testeandroidv2.model
+package com.example.testeandroidv2.domain.login
 
 
 import com.squareup.moshi.Json
