+++ App/app/src/main/java/com/example/testeandroidv2/domain/statements/Statement.kt
@@ -1,4 +1,4 @@
-package com.example.testeandroidv2.model
+package com.example.testeandroidv2.domain.statements
 
 
 import com.squareup.moshi.Json
