+++ App/app/src/main/java/com/example/testeandroidv2/domain/statements/StatementsBodyResponse.kt
@@ -1,5 +1,7 @@
-package com.example.testeandroidv2.model
+package com.example.testeandroidv2.domain.statements
 
+import com.example.testeandroidv2.domain.statements.ErrorStatement
+import com.example.testeandroidv2.domain.statements.Statement
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
